[refactor] 쿠폰 구조 변경 및 repo 분리
