+++ src/main/java/kr/hhplus/be/server/domain/coupon/presenter/controller/dto/IssueCouponResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.coupon.present.controller.dto;
+package kr.hhplus.be.server.domain.coupon.presenter.controller.dto;
 
 import kr.hhplus.be.server.domain.coupon.application.service.dto.IssueCouponResult;
 import lombok.Data;
