+++ src/main/java/kr/hhplus/be/server/domain/product/application/facade/ProductFacade.java
@@ -1,9 +1,9 @@
-package kr.hhplus.be.server.domain.product.facade;
+package kr.hhplus.be.server.domain.product.application.facade;
 
 import kr.hhplus.be.server.domain.order.service.OrderProductService;
 import kr.hhplus.be.server.domain.payment.application.service.PaymentService;
-import kr.hhplus.be.server.domain.product.entity.Product;
-import kr.hhplus.be.server.domain.product.service.ProductService;
+import kr.hhplus.be.server.domain.product.domain.entity.Product;
+import kr.hhplus.be.server.domain.product.application.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
