+++ src/main/java/kr/hhplus/be/server/domain/product/infra/ProductStockJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.domain.product.infra;
+
+import kr.hhplus.be.server.domain.product.domain.entity.ProductStock;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductStockJpaRepository extends JpaRepository<ProductStock, Long> {
+}
