--- src/main/java/kr/hhplus/be/server/domain/product/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.domain.product.repository;
-
-import kr.hhplus.be.server.domain.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
