--- src/main/java/kr/hhplus/be/server/domain/product/repository/ProductStockRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.domain.product.repository;
-
-import kr.hhplus.be.server.domain.product.entity.ProductStock;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductStockRepository extends JpaRepository<ProductStock, Long> {
-}
