+++ src/test/java/kr/hhplus/be/server/domain/product/controller/ProductControllerIntegrationTest.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.domain.product.controller;
 
-import kr.hhplus.be.server.domain.product.facade.ProductFacade;
-import kr.hhplus.be.server.domain.product.service.ProductService;
+import kr.hhplus.be.server.domain.product.application.facade.ProductFacade;
+import kr.hhplus.be.server.domain.product.application.service.ProductService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
