+++ src/test/java/kr/hhplus/be/server/domain/product/facade/ProductFacadeIntegrationTest.java
@@ -1,6 +1,7 @@
 package kr.hhplus.be.server.domain.product.facade;
 
-import kr.hhplus.be.server.domain.product.entity.Product;
+import kr.hhplus.be.server.domain.product.application.facade.ProductFacade;
+import kr.hhplus.be.server.domain.product.domain.entity.Product;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
